Phone and Email Search from a clipboard
If you are task with to find phone numbers and email addresses from a vast library or any webpage. This little automation with python can help you search and paste only phone numbers and email addresses from the clipboard. 

All you have to do is Ctrl + A to select from the source and Ctrl + C to copy to the clipboard. Then run your program to search only phone numbers and Email addresses. 

To know more check out the book - Automate the Boring Stuff with Python